// src/app/models/extrait.ts
export interface Extrait {
   id: String;
   pouranneec: Date;
   pouranneel: String;
   numregistrec: number;
   numregistrel: String;
   datenaissancel: String;
   lieunaissance: String;
   prenom: String;
   nom: String;
   prenomp: String;
   prenomm: String;
   nomm: String;
   sexe: String;
   datedelivrancel: String;
   datedelivrancec: Date;   
};

export interface Residence {
   id: String;
   prenom: String;
   nom: String;
   datenaissance:Date;
   lieuresidence: String;
   quartier: String;
   dure: String;
   lieunaissance: String;
  
};